requirements.txt
django>=5.2,<6.0
channels>=4.2,<5.0
daphne>=4.1,<5.0
django-cors-headers>=4.7,<5.0
bcrypt>=4.2,<5.0
paho-mqtt>=2.1,<3.0
